Add ability to build python modules/processors.
Checking to see if we can build against ROOTs python.

Adding Bindings as a temporary commit.

Making updates to the data processor to build in the new configure schema.

Update the .gitignore for Bindings

Adding ability to build Bindings directory

Exanding out the Bindings classes and adding ORPyTreeWriter

Removing verbose output during cleaning stage.

Adding ORPyDataProcManager to ensure we may link against ORManagement.

Fixing the build to not require all symbols to be resolved in the libraries.

Removing StartRun and EndRun as this overloads base functions.  Including GetTree and DoNotAutoFillTree to export the interface to python.

Being more careful about checking for Py_None, which indicates that the functions are not overloaded.

Using PyBufferFactory to send buffers.

Adding a configure check to ensure >= 5.34 to build bindings.

Updating READMEs

Fix README to deal with array syntax.

Allowing undefined symbols to be handled at runtime on Mac OS X.

Switching order of header files, there's an issue with the Python headers on Mac OS X.

Removing requirement on PyBufferFactory.

Checking for ROOT 5.34.02…

configure updates.

git-svn-id: svn://orca.physics.unc.edu/OrcaRoot@9081 fd86f90f-2c0a-422a-b972-84106d8db31b